tests/test_case_upstream_downstream_pipeline/infrastructure_code/requirements/requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
